+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 New:
 - Added support for DataSets
-- Added new Device rename command
+- Added new Device set_name command
 - Redesigned Broker configuration UI
 - Added support for manual connection of the EoN
 - Added connect command for EoN
-- Added rename command for EoN node
+- Added set_name command for EoN node
 
 ### 1.4.1: Maintenance Release